# [tsdb] Improve mergeSeriesSet