Using the idea similar to [tsdb/pr616](https://github.com/prometheus/tsdb/pull/616) to improve both time and alloc.
Currently I use the original `mergeSeriesSet` if the len of the passed-in seriesset slice = **2**, because I think the original one is the most efficient way for merging only **2** seriesset.
Benchmark results of `BenchmarkMergedSeriesSet`.
```
benchmark                                             old ns/op     new ns/op     delta
BenchmarkMergedSeriesSet/series=100,blocks=1-8        4746          4807          +1.29%
BenchmarkMergedSeriesSet/series=100,blocks=2-8        43989         44127         +0.31%
BenchmarkMergedSeriesSet/series=100,blocks=4-8        130592        89924         -31.14%
BenchmarkMergedSeriesSet/series=100,blocks=8-8        291087        165807        -43.04%
BenchmarkMergedSeriesSet/series=100,blocks=16-8       653520        331504        -49.27%
BenchmarkMergedSeriesSet/series=100,blocks=32-8       1344424       660384        -50.88%
BenchmarkMergedSeriesSet/series=1000,blocks=1-8       8172          8902          +8.93%
BenchmarkMergedSeriesSet/series=1000,blocks=2-8       441894        443515        +0.37%
BenchmarkMergedSeriesSet/series=1000,blocks=4-8       1311465       939890        -28.33%
BenchmarkMergedSeriesSet/series=1000,blocks=8-8       3088490       1883928       -39.00%
BenchmarkMergedSeriesSet/series=1000,blocks=16-8      6871918       3531934       -48.60%
BenchmarkMergedSeriesSet/series=1000,blocks=32-8      12685259      6169894       -51.36%
BenchmarkMergedSeriesSet/series=10000,blocks=1-8      36889         36654         -0.64%
BenchmarkMergedSeriesSet/series=10000,blocks=2-8      3186629       3136228       -1.58%
BenchmarkMergedSeriesSet/series=10000,blocks=4-8      8641566       6963987       -19.41%
BenchmarkMergedSeriesSet/series=10000,blocks=8-8      18862093      13080495      -30.65%
BenchmarkMergedSeriesSet/series=10000,blocks=16-8     41348147      24343738      -41.12%
BenchmarkMergedSeriesSet/series=10000,blocks=32-8     86234288      48789802      -43.42%
BenchmarkMergedSeriesSet/series=20000,blocks=1-8      69872         68857         -1.45%
BenchmarkMergedSeriesSet/series=20000,blocks=2-8      7255853       7572243       +4.36%
BenchmarkMergedSeriesSet/series=20000,blocks=4-8      17507968      15065738      -13.95%
BenchmarkMergedSeriesSet/series=20000,blocks=8-8      37290957      26177634      -29.80%
BenchmarkMergedSeriesSet/series=20000,blocks=16-8     78885648      48284644      -38.79%
BenchmarkMergedSeriesSet/series=20000,blocks=32-8     173633339     98093248      -43.51%

benchmark                                             old allocs     new allocs     delta
BenchmarkMergedSeriesSet/series=100,blocks=1-8        11             11             +0.00%
BenchmarkMergedSeriesSet/series=100,blocks=2-8        217            217            +0.00%
BenchmarkMergedSeriesSet/series=100,blocks=4-8        628            232            -63.06%
BenchmarkMergedSeriesSet/series=100,blocks=8-8        1449           251            -82.68%
BenchmarkMergedSeriesSet/series=100,blocks=16-8       3090           286            -90.74%
BenchmarkMergedSeriesSet/series=100,blocks=32-8       6371           353            -94.46%
BenchmarkMergedSeriesSet/series=1000,blocks=1-8       11             11             +0.00%
BenchmarkMergedSeriesSet/series=1000,blocks=2-8       2017           2017           +0.00%
BenchmarkMergedSeriesSet/series=1000,blocks=4-8       6028           2032           -66.29%
BenchmarkMergedSeriesSet/series=1000,blocks=8-8       14049          2051           -85.40%
BenchmarkMergedSeriesSet/series=1000,blocks=16-8      30090          2086           -93.07%
BenchmarkMergedSeriesSet/series=1000,blocks=32-8      62171          2153           -96.54%
BenchmarkMergedSeriesSet/series=10000,blocks=1-8      11             11             +0.00%
BenchmarkMergedSeriesSet/series=10000,blocks=2-8      20017          20017          +0.00%
BenchmarkMergedSeriesSet/series=10000,blocks=4-8      60028          20032          -66.63%
BenchmarkMergedSeriesSet/series=10000,blocks=8-8      140049         20051          -85.68%
BenchmarkMergedSeriesSet/series=10000,blocks=16-8     300090         20086          -93.31%
BenchmarkMergedSeriesSet/series=10000,blocks=32-8     620171         20153          -96.75%
BenchmarkMergedSeriesSet/series=20000,blocks=1-8      11             11             +0.00%
BenchmarkMergedSeriesSet/series=20000,blocks=2-8      40017          40017          +0.00%
BenchmarkMergedSeriesSet/series=20000,blocks=4-8      120028         40032          -66.65%
BenchmarkMergedSeriesSet/series=20000,blocks=8-8      280049         40051          -85.70%
BenchmarkMergedSeriesSet/series=20000,blocks=16-8     600090         40086          -93.32%
BenchmarkMergedSeriesSet/series=20000,blocks=32-8     1240171        40153          -96.76%

benchmark                                             old bytes     new bytes     delta
BenchmarkMergedSeriesSet/series=100,blocks=1-8        552           552           +0.00%
BenchmarkMergedSeriesSet/series=100,blocks=2-8        7184          7184          +0.00%
BenchmarkMergedSeriesSet/series=100,blocks=4-8        20448         10920         -46.60%
BenchmarkMergedSeriesSet/series=100,blocks=8-8        46976         18184         -61.29%
BenchmarkMergedSeriesSet/series=100,blocks=16-8       100032        32712         -67.30%
BenchmarkMergedSeriesSet/series=100,blocks=32-8       206144        61768         -70.04%
BenchmarkMergedSeriesSet/series=1000,blocks=1-8       552           552           +0.00%
BenchmarkMergedSeriesSet/series=1000,blocks=2-8       64784         64784         +0.00%
BenchmarkMergedSeriesSet/series=1000,blocks=4-8       193248        97320         -49.64%
BenchmarkMergedSeriesSet/series=1000,blocks=8-8       450176        162184        -63.97%
BenchmarkMergedSeriesSet/series=1000,blocks=16-8      964033        291912        -69.72%
BenchmarkMergedSeriesSet/series=1000,blocks=32-8      1991745       551368        -72.32%
BenchmarkMergedSeriesSet/series=10000,blocks=1-8      552           552           +0.00%
BenchmarkMergedSeriesSet/series=10000,blocks=2-8      640784        640784        +0.00%
BenchmarkMergedSeriesSet/series=10000,blocks=4-8      1921249       961320        -49.96%
BenchmarkMergedSeriesSet/series=10000,blocks=8-8      4482178       1602185       -64.25%
BenchmarkMergedSeriesSet/series=10000,blocks=16-8     9604032       2883913       -69.97%
BenchmarkMergedSeriesSet/series=10000,blocks=32-8     19847753      5447371       -72.55%
BenchmarkMergedSeriesSet/series=20000,blocks=1-8      552           552           +0.00%
BenchmarkMergedSeriesSet/series=20000,blocks=2-8      1280784       1280784       +0.00%
BenchmarkMergedSeriesSet/series=20000,blocks=4-8      3841249       1921320       -49.98%
BenchmarkMergedSeriesSet/series=20000,blocks=8-8      8962178       3202184       -64.27%
BenchmarkMergedSeriesSet/series=20000,blocks=16-8     19204032      5763912       -69.99%
BenchmarkMergedSeriesSet/series=20000,blocks=32-8     39687748      10887368      -72.57%
```