@krasi-georgiev 2019-08-29
I used the PR to test the new automated bench testing

here are the results:
https://github.com/prometheus-community/prometheus/pull/7#issuecomment-526263619

@brian-brazil 	should be able to comment about the PR changes.