@naivewong 2019-08-30
@krasi-georgiev  I ran the benchmark on the prometheus machine just now. Actually I found the difference is under 5% for all cases in `BenchmarkRangeQuery`.