@krasi-georgiev 2019-08-30
yes you are right, I completely forgot that the GH actions are run in shared VM env so the results will be inconsistent. Will refactor the testing to use dedicated machines.