@brian-brazil 2019-09-02
This looks good. Do benchmarks show much of a difference for two series sets to justify the separate handling?