@naivewong 2019-09-03
The benchmarks results on the op time seem a bit better.
```benchmark                                              old ns/op     new ns/op     delta
BenchmarkMergedSeriesSet/series=100,blocks=1-48        4806          4828          +0.46%
BenchmarkMergedSeriesSet/series=100,blocks=2-48        36804         35332         -4.00%
BenchmarkMergedSeriesSet/series=100,blocks=4-48        103204        79611         -22.86%
BenchmarkMergedSeriesSet/series=100,blocks=8-48        248132        156220        -37.04%
BenchmarkMergedSeriesSet/series=100,blocks=16-48       544840        311781        -42.78%
BenchmarkMergedSeriesSet/series=100,blocks=32-48       1179173       620354        -47.39%
BenchmarkMergedSeriesSet/series=1000,blocks=1-48       10913         10851         -0.57%
BenchmarkMergedSeriesSet/series=1000,blocks=2-48       331775        325112        -2.01%
BenchmarkMergedSeriesSet/series=1000,blocks=4-48       1032389       765167        -25.88%
BenchmarkMergedSeriesSet/series=1000,blocks=8-48       2530947       1543987       -39.00%
BenchmarkMergedSeriesSet/series=1000,blocks=16-48      5705631       3109198       -45.51%
BenchmarkMergedSeriesSet/series=1000,blocks=32-48      12413230      6263710       -49.54%
BenchmarkMergedSeriesSet/series=10000,blocks=1-48      72075         71672         -0.56%
BenchmarkMergedSeriesSet/series=10000,blocks=2-48      3707168       3659727       -1.28%
BenchmarkMergedSeriesSet/series=10000,blocks=4-48      10936970      8235542       -24.70%
BenchmarkMergedSeriesSet/series=10000,blocks=8-48      26332919      16343411      -37.94%
BenchmarkMergedSeriesSet/series=10000,blocks=16-48     58437574      32342400      -44.65%
BenchmarkMergedSeriesSet/series=10000,blocks=32-48     130768647     65279577      -50.08%
BenchmarkMergedSeriesSet/series=20000,blocks=1-48      139307        139256        -0.04%
BenchmarkMergedSeriesSet/series=20000,blocks=2-48      8976742       9290798       +3.50%
BenchmarkMergedSeriesSet/series=20000,blocks=4-48      23147352      17791593      -23.14%
BenchmarkMergedSeriesSet/series=20000,blocks=8-48      54680802      33260669      -39.17%
BenchmarkMergedSeriesSet/series=20000,blocks=16-48     120038946     65521638      -45.42%
BenchmarkMergedSeriesSet/series=20000,blocks=32-48     261474153     128783898     -50.75%

benchmark                                              old allocs     new allocs     delta
BenchmarkMergedSeriesSet/series=100,blocks=1-48        11             11             +0.00%
BenchmarkMergedSeriesSet/series=100,blocks=2-48        217            217            +0.00%
BenchmarkMergedSeriesSet/series=100,blocks=4-48        628            232            -63.06%
BenchmarkMergedSeriesSet/series=100,blocks=8-48        1449           251            -82.68%
BenchmarkMergedSeriesSet/series=100,blocks=16-48       3090           286            -90.74%
BenchmarkMergedSeriesSet/series=100,blocks=32-48       6371           353            -94.46%
BenchmarkMergedSeriesSet/series=1000,blocks=1-48       11             11             +0.00%
BenchmarkMergedSeriesSet/series=1000,blocks=2-48       2017           2017           +0.00%
BenchmarkMergedSeriesSet/series=1000,blocks=4-48       6028           2032           -66.29%
BenchmarkMergedSeriesSet/series=1000,blocks=8-48       14049          2051           -85.40%
BenchmarkMergedSeriesSet/series=1000,blocks=16-48      30090          2086           -93.07%
BenchmarkMergedSeriesSet/series=1000,blocks=32-48      62171          2153           -96.54%
BenchmarkMergedSeriesSet/series=10000,blocks=1-48      11             11             +0.00%
BenchmarkMergedSeriesSet/series=10000,blocks=2-48      20017          20017          +0.00%
BenchmarkMergedSeriesSet/series=10000,blocks=4-48      60028          20032          -66.63%
BenchmarkMergedSeriesSet/series=10000,blocks=8-48      140049         20051          -85.68%
BenchmarkMergedSeriesSet/series=10000,blocks=16-48     300090         20086          -93.31%
BenchmarkMergedSeriesSet/series=10000,blocks=32-48     620171         20153          -96.75%
BenchmarkMergedSeriesSet/series=20000,blocks=1-48      11             11             +0.00%
BenchmarkMergedSeriesSet/series=20000,blocks=2-48      40017          40017          +0.00%
BenchmarkMergedSeriesSet/series=20000,blocks=4-48      120028         40032          -66.65%
BenchmarkMergedSeriesSet/series=20000,blocks=8-48      280049         40051          -85.70%
BenchmarkMergedSeriesSet/series=20000,blocks=16-48     600090         40086          -93.32%
BenchmarkMergedSeriesSet/series=20000,blocks=32-48     1240171        40153          -96.76%

benchmark                                              old bytes     new bytes     delta
BenchmarkMergedSeriesSet/series=100,blocks=1-48        552           552           +0.00%
BenchmarkMergedSeriesSet/series=100,blocks=2-48        7184          7184          +0.00%
BenchmarkMergedSeriesSet/series=100,blocks=4-48        20448         10936         -46.52%
BenchmarkMergedSeriesSet/series=100,blocks=8-48        46976         18200         -61.26%
BenchmarkMergedSeriesSet/series=100,blocks=16-48       100032        32728         -67.28%
BenchmarkMergedSeriesSet/series=100,blocks=32-48       206146        61785         -70.03%
BenchmarkMergedSeriesSet/series=1000,blocks=1-48       552           552           +0.00%
BenchmarkMergedSeriesSet/series=1000,blocks=2-48       64784         64784         +0.00%
BenchmarkMergedSeriesSet/series=1000,blocks=4-48       193250        97337         -49.63%
BenchmarkMergedSeriesSet/series=1000,blocks=8-48       450182        162201        -63.97%
BenchmarkMergedSeriesSet/series=1000,blocks=16-48      964036        291929        -69.72%
BenchmarkMergedSeriesSet/series=1000,blocks=32-48      1991756       551386        -72.32%
BenchmarkMergedSeriesSet/series=10000,blocks=1-48      552           552           +0.00%
BenchmarkMergedSeriesSet/series=10000,blocks=2-48      640784        640784        +0.00%
BenchmarkMergedSeriesSet/series=10000,blocks=4-48      1921251       961337        -49.96%
BenchmarkMergedSeriesSet/series=10000,blocks=8-48      4482179       1602201       -64.25%
BenchmarkMergedSeriesSet/series=10000,blocks=16-48     9604055       2883928       -69.97%
BenchmarkMergedSeriesSet/series=10000,blocks=32-48     19847750      5447391       -72.55%
BenchmarkMergedSeriesSet/series=20000,blocks=1-48      552           552           +0.00%
BenchmarkMergedSeriesSet/series=20000,blocks=2-48      1280784       1280784       +0.00%
BenchmarkMergedSeriesSet/series=20000,blocks=4-48      3841249       1921336       -49.98%
BenchmarkMergedSeriesSet/series=20000,blocks=8-48      8962182       3202200       -64.27%
BenchmarkMergedSeriesSet/series=20000,blocks=16-48     19204039      5763930       -69.99%
BenchmarkMergedSeriesSet/series=20000,blocks=32-48     39687744      10887384      -72.57%