@brian-brazil 2019-09-03
What's that comparing? It looks largely in the noise.