@naivewong 2019-11-08
Sorry, I just realized I misunderstood what Brian said before. I thought he just asked me to compare the original and the current method for two series. As a result, the previous result did not show the results without the binaryMergedSeriesSet for 2 series. The following is the actual results without binaryMergedSeriesSet for 2 series -
```
benchmark                                            old ns/op     new ns/op     delta
BenchmarkMergedSeriesSet/series=100,blocks=2-8       43791         48534         +10.83%
BenchmarkMergedSeriesSet/series=1000,blocks=2-8      478730        510211        +6.58%
BenchmarkMergedSeriesSet/series=10000,blocks=2-8     3509296       3685314       +5.02%
BenchmarkMergedSeriesSet/series=20000,blocks=2-8     8970934       8713768       -2.87%
```
That's why I added binaryMergedSeriesSet. Sorry for the mistake again.